+++ packages/visual-stack-docs/src/containers/Components/Docs/badge.js
@@ -2,12 +2,12 @@ import React from 'react';
 import { Panel, Body } from '@cjdev/visual-stack/lib/components/Panel';
 import { Demo, Snippet } from '../../../components/Demo';
 /* s2:start */
-import { Chip } from '@cjdev/visual-stack';
+import { Badge } from '@cjdev/visual-stack';
 /* s2:end */
 import Box from '@cjdev/visual-stack/lib/experimental/Box';
 
 export default () => (
-  <Demo srcFile="/samples/src/containers/Components/Docs/chip.js">
+  <Demo srcFile="/samples/src/containers/Components/Docs/Badge.js">
     {snippets => {
       return (
         <div>
@@ -15,11 +15,11 @@ export default () => (
             <Body>
               {/* s1:start */}
               <Box direction="row" gap="default">
-                <Chip backgroundColor="#9C52F7">CJ Performer</Chip>
-                <Chip backgroundColor="#EA4B93">CJ Spotlight</Chip>
-                <Chip backgroundColor="#4574E2">Content Certified</Chip>
-                <Chip backgroundColor="#0CAAAC">New to Network</Chip>
-                <Chip backgroundColor="#072477">SubAffiliate</Chip>
+                <Badge backgroundColor="#9C52F7">CJ Performer</Badge>
+                <Badge backgroundColor="#EA4B93">CJ Spotlight</Badge>
+                <Badge backgroundColor="#4574E2">Content Certified</Badge>
+                <Badge backgroundColor="#0CAAAC">New to Network</Badge>
+                <Badge backgroundColor="#072477">SubAffiliate</Badge>
               </Box>
               {/* s1:end */}
               <Snippet tag="s2" src={snippets} />